+++ src/ScanRunners.js
@@ -2,7 +2,7 @@ import Scanner from './Scanner';
 
 
 // Scan initiators
-const ScanRunners = function (socket, extensionName, validatorsGetter) {
+const ScanRunners = function (socket, extensionName, configGetter) {
 	const reduceGroupedPath = (reduced, info) => {
 		reduced.push(...info.paths);
 		return reduced;
@@ -15,12 +15,46 @@ const ScanRunners = function (socket, extensionName, validatorsGetter) {
 		});
 	};
 
+	// Log extension info (debug) message after the scan was completed
+	const logCompleted = (scanner, message) => {
+		let text = message;
+		text += ` (took ${scanner.stats.duration} ms`;
+		if (scanner.stats.ignoredFiles > 0 || scanner.stats.ignoredDirectories > 0) {
+			text += `, ignored ${scanner.stats.ignoredDirectories} directories and ${scanner.stats.ignoredFiles} files`;
+		}
+		text += ')';
+
+		socket.logger.info(text);
+	};
+
+	const pathValidator = (skipQueueCheck) => {
+		if (!configGetter().ignoreExcluded) {
+			return () => true;
+		}
+
+		const validate = async (path) => {
+			try {
+				await socket.post('share/validate_path', {
+					path,
+					skip_queue_check: skipQueueCheck,
+				});
+			} catch (e) {
+				socket.logger.verbose(`Path ${path} is ignored from share`);
+				return false;
+			}
+
+			return true;
+		};
+
+		return validate;
+	};
+
 	const errorLogger = message => {
 		// TODO: add file logger support
 		postEvent(message, 'warning');
 	};
 
-	const onScanCompleted = (scanner) => {
+	const onManualScanCompleted = (scanner) => {
 		let text;
 		if (scanner.errors.count()) {
 			text = `Scan completed and the following problems were found: ${scanner.errors.format()}`;
@@ -28,7 +62,7 @@ const ScanRunners = function (socket, extensionName, validatorsGetter) {
 			text = 'Scan completed, no problems were found';
 		}
 
-		socket.logger.info(`Scan completed: ${scanner.stats.scanned} paths scanned with maximum concurrency of ${scanner.stats.maxRunning}`);
+		logCompleted(scanner, `Scan completed: ${scanner.stats.scanned} paths scanned with maximum concurrency of ${scanner.stats.maxRunning}`);
 		postEvent(text, scanner.errors.count() ? 'warning' : 'info');
 	};
 
@@ -37,9 +71,9 @@ const ScanRunners = function (socket, extensionName, validatorsGetter) {
 		const text = paths.length === 1 ? `Scanning the path ${paths[0]}...` : `Scanning ${paths.length} paths...`;
 		postEvent(text, 'info');
 
-		const scanner = Scanner(validatorsGetter(), errorLogger);
+		const scanner = Scanner(configGetter().validators, errorLogger, pathValidator(false));
 		await scanner.scanPaths(paths);
-		onScanCompleted(scanner);
+		onManualScanCompleted(scanner);
 		
 		return scanner;
 	};
@@ -49,9 +83,9 @@ const ScanRunners = function (socket, extensionName, validatorsGetter) {
 		const directories = await socket.get('share/grouped_root_paths');
 
 		postEvent('Scanning shared releases...', 'info');
-		const scanner = Scanner(validatorsGetter(), errorLogger);
+		const scanner = Scanner(configGetter().validators, errorLogger, pathValidator(false));
 		await scanner.scanPaths(directories.reduce(reduceGroupedPath, []));
-		onScanCompleted(scanner);
+		onManualScanCompleted(scanner);
 
 		return scanner;
 	};
@@ -64,10 +98,10 @@ const ScanRunners = function (socket, extensionName, validatorsGetter) {
 		}
 
 		// Scan it
-		const scanner = Scanner(validatorsGetter(), errorLogger);
+		const scanner = Scanner(configGetter().validators, errorLogger, pathValidator(true));
 		await scanner.scanPath(bundle.target);
 
-		socket.logger.info(`Bundle scan completed: ${scanner.stats.scanned} paths were scanned`);
+		logCompleted(scanner, `Bundle scan completed: ${scanner.stats.scanned} paths were scanned`);
 		if (scanner.errors.count()) {
 			// Failed, report and reject
 			const error = scanner.errors.pickOne();
@@ -88,10 +122,10 @@ const ScanRunners = function (socket, extensionName, validatorsGetter) {
 	// Scan new share directories
 	const onShareDirectoryAdded = async ({ path, new_parent }, accept, reject) => {
 		// Scan it
-		const scanner = Scanner(validatorsGetter(), errorLogger);
+		const scanner = Scanner(configGetter().validators, errorLogger, pathValidator(false));
 		await scanner.scanPath(path, false);
 
-		socket.logger.info(`Share directory scan completed: ${scanner.stats.scanned} paths were scanned`);
+		logCompleted(scanner, `Share directory scan completed: ${scanner.stats.scanned} paths were scanned`);
 		if (scanner.errors.count()) {
 			// Failed, report and reject
 			const error = scanner.errors.pickOne();