+++ tests/ScanRunners.test.js
@@ -11,6 +11,18 @@ describe('Scan runner', () => {
 		error: _ => {},
 	}
 
+	const getScanRunners = (socket, ignoreExcluded = false) => {
+		return ScanRunners(
+			socket, 
+			'test-extension',
+			() => ({
+				validators,
+				ignoreExcluded,
+			}),
+			
+		);
+	};
+
 	test('should reject invalid bundles', async () => {
 		const scanPath = path.join(__dirname, 'data/Test.Release-TEST');
 
@@ -30,11 +42,13 @@ describe('Scan runner', () => {
 		const reject = jest.fn();
 		const accept = jest.fn();
 
-		const runner = ScanRunners(socket, 'test-extension', _ => validators);
-		await runner.onBundleFinished(bundle, accept, reject);
+		const runner = getScanRunners(socket);
+		const scanner = await runner.onBundleFinished(bundle, accept, reject);
 
 		expect(reject.mock.calls.length).toBe(1);
 		expect(accept.mock.calls.length).toBe(0);
+		expect(accept.mock.calls.length).toBe(0);
+		expect(scanner.stats.scanned).toBe(2);
 	});
 
 	test('should reject invalid new share directories', async () => {
@@ -53,11 +67,55 @@ describe('Scan runner', () => {
 		const reject = jest.fn();
 		const accept = jest.fn();
 
-		const runner = ScanRunners(socket, 'test-extension', _ => validators);
-		await runner.onShareDirectoryAdded(hookData, accept, reject);
+		const runner = getScanRunners(socket);
+		const scanner = await runner.onShareDirectoryAdded(hookData, accept, reject);
 
 		expect(reject.mock.calls.length).toBe(1);
 		expect(accept.mock.calls.length).toBe(0);
+		expect(scanner.stats.scanned).toBe(1); // Not recursive
+	});
+
+	test('should ignore excluded files/directories', async () => {
+		const scanPath = path.join(__dirname, 'data/Test.Release-TEST');
+
+		const ignoredPathFn = jest.fn();
+		const socket = {
+			post: (url, data) => {
+				// Events
+				if (url.startsWith('events')) {
+					return {};
+				}
+
+				if (url.startsWith('share')) {
+
+					// Share validator
+					if (data.path.endsWith('Sample' + path.sep) || data.path.endsWith('forbidden_extra.zip')) {
+						ignoredPathFn(url, data);
+						throw Error('Ignored');
+					}
+				}
+
+				return {};
+			},
+			logger,
+		};
+
+		const hookData = {
+			path: scanPath,
+			new_parent: false,
+		};
+
+		
+		const reject = jest.fn();
+		const accept = jest.fn();
+
+		const runner = getScanRunners(socket, true);
+		const scanner = await runner.onShareDirectoryAdded(hookData, accept, reject);
+
+		expect(reject.mock.calls.length).toBe(0);
+		expect(accept.mock.calls.length).toBe(1);
+		expect(ignoredPathFn.mock.calls.length).toBe(2);
+		expect(scanner.stats.scanned).toBe(1);
 	});
 
 	test('should scan share roots', async () => {
@@ -74,10 +132,11 @@ describe('Scan runner', () => {
 			logger,
 		};
 
-		const runner = ScanRunners(socket, 'test-extension', _ => validators);
+		const runner = getScanRunners(socket);
 		const scanner = await runner.scanShareRoots([ 1 ]);
 
 		expect(scanner.errors.count() > 0).toBe(true);
+		expect(scanner.stats.scanned).toBe(2);
 	});
 
 	test('should perform share scan', async () => {
@@ -94,7 +153,7 @@ describe('Scan runner', () => {
 			logger,
 		};
 
-		const runner = ScanRunners(socket, 'test-extension', _ => validators);
+		const runner = getScanRunners(socket);
 		const scanner = await runner.scanShare();
 
 		expect(scanner.errors.count() > 0).toBe(true);
@@ -114,7 +173,7 @@ describe('Scan runner', () => {
 			logger,
 		};
 
-		const runner = ScanRunners(socket, 'test-extension', _ => validators);
+		const runner = getScanRunners(socket);
 		const scanner = await runner.scanShare();
 
 		expect(scanner.errors.count() > 0).toBe(true);