+++ tests/validators/ExtraNfoSfv.test.js
@@ -1,15 +1,13 @@
 import ExtraNfoSfv from 'validators/ExtraNfoSfv';
 import path from 'path';
-import Scanner from 'Scanner';
+import { getTestScanner } from '../helpers';
 
 
 describe('Missing NFO/SFV validator', () => {
-	const errorLogger = error => console.log(error);
-
 	test('should detect duplicate NFO and SFV files', async () => {
 		const scanPath = path.join(__dirname, 'ExtraNfoSfv-data/Duplicate.NFO.SFV-TEST');
 
-		const scanner = Scanner([ ExtraNfoSfv ], errorLogger);
+		const scanner = getTestScanner([ ExtraNfoSfv ]);
 		await scanner.scanPath(scanPath);
 
 		expect(scanner.errors.count('multiple_nfo_files')).toEqual(1);
@@ -20,7 +18,7 @@ describe('Missing NFO/SFV validator', () => {
 	test('should detect directories with no release files', async () => {
 		const scanPath = path.join(__dirname, 'ExtraNfoSfv-data/Release.Files.Missing-TEST');
 
-		const scanner = Scanner([ ExtraNfoSfv ], errorLogger);
+		const scanner = getTestScanner([ ExtraNfoSfv ]);
 		await scanner.scanPath(scanPath);
 
 		expect(scanner.errors.count('no_release_files')).toEqual(1);
@@ -31,7 +29,7 @@ describe('Missing NFO/SFV validator', () => {
 	test('should not detect nfo fix directories', async () => {
 		const scanPath = path.join(__dirname, 'ExtraNfoSfv-data/Test.NFO.FIX-TEST');
 
-		const scanner = Scanner([ ExtraNfoSfv ], errorLogger);
+		const scanner = getTestScanner([ ExtraNfoSfv ]);
 		scanner.scanPath(scanPath);
 
 		expect(scanner.errors.count('no_release_files')).toEqual(0);