+++ tests/validators/MissingNfo.test.js
@@ -1,15 +1,13 @@
 import MissingNfo from 'validators/MissingNfo';
 import path from 'path';
-import Scanner from 'Scanner';
+import { getTestScanner } from '../helpers';
 
 
 describe('Missing NFO validator', () => {
-	const errorLogger = error => console.log(error);
-
 	test('should detect missing NFO files', async () => {
 		const scanPath = path.join(__dirname, 'MissingNfo-data/Missing.NFO-TEST');
 
-		const scanner = Scanner([ MissingNfo ], errorLogger);
+		const scanner = getTestScanner([ MissingNfo ]);
 		await scanner.scanPath(scanPath);
 
 		expect(scanner.errors.count('nfo_missing')).toEqual(1);
@@ -19,7 +17,7 @@ describe('Missing NFO validator', () => {
 		const scanPath = path.join(__dirname, 'MissingNfo-data/NFO.Subdirectory-Test');
 		//const scanPath = 'C:\\Program Files (x86)\\Microsoft Visual Studio 14.0\\';
 
-		const scanner = Scanner([ MissingNfo ], errorLogger);
+		const scanner = getTestScanner([ MissingNfo ]);
 		await scanner.scanPath(scanPath);
 
 		expect(scanner.errors.count('nfo_missing')).toEqual(0);
@@ -28,7 +26,7 @@ describe('Missing NFO validator', () => {
 	test('should not report non-release directories', async () => {
 		const scanPath = path.join(__dirname, 'MissingNfo-data/false.positive-test');
 
-		const scanner = Scanner([ MissingNfo ], errorLogger);
+		const scanner = getTestScanner([ MissingNfo ]);
 		await scanner.scanPath(scanPath);
 
 		expect(scanner.errors.count('nfo_missing')).toEqual(0);
@@ -37,7 +35,7 @@ describe('Missing NFO validator', () => {
 	test('should handle release subdirectories', async () => {
 		const scanPath = path.join(__dirname, 'MissingNfo-data/Release.Subdirectory-Test');
 
-		const scanner = Scanner([ MissingNfo ], errorLogger);
+		const scanner = getTestScanner([ MissingNfo ]);
 		await scanner.scanPath(scanPath);
 
 		expect(scanner.errors.count('nfo_missing')).toEqual(1);