+++ tests/validators/SFVChecker.test.js
@@ -1,15 +1,13 @@
 import SFVChecker from 'validators/SFVChecker';
 import path from 'path';
-import Scanner from 'Scanner';
+import { getTestScanner } from '../helpers';
 
 
 describe('SFV checker', () => {
-	const errorLogger = error => console.log(error);
-
 	test('should detect missing and extra files', async () => {
 		const scanPath = path.join(__dirname, 'SFVChecker-data/Test.Release-TEST');
 
-		const scanner = Scanner([ SFVChecker ], errorLogger);
+		const scanner = getTestScanner([ SFVChecker ]);
 		await scanner.scanPath(scanPath);
 
 		expect(scanner.errors.count('file_missing')).toEqual(1);
@@ -20,7 +18,7 @@ describe('SFV checker', () => {
 	test('should detect invalid SFV files', async () => {
 		const scanPath = path.join(__dirname, 'SFVChecker-data/Empty.SFV-TEST');
 
-		const scanner = Scanner([ SFVChecker ], errorLogger);
+		const scanner = getTestScanner([ SFVChecker ]);
 		await scanner.scanPath(scanPath);
 
 		expect(scanner.errors.count('file_missing')).toEqual(0);
@@ -31,7 +29,7 @@ describe('SFV checker', () => {
 	test('should not detect audiobook extras', async () => {
 		const scanPath = path.join(__dirname, 'SFVChecker-data/Audiobook-(2005)-AUDIOBOOK-2006-TEST');
 
-		const scanner = Scanner([ SFVChecker ], errorLogger);
+		const scanner = getTestScanner([ SFVChecker ]);
 		await scanner.scanPath(scanPath);
 
 		expect(scanner.errors.count()).toEqual(0);
@@ -40,7 +38,7 @@ describe('SFV checker', () => {
 	test('should perform scan if no release files are present', async () => {
 		const scanPath = path.join(__dirname, 'SFVChecker-data/No.Release.Files-TEST');
 
-		const scanner = Scanner([ SFVChecker ], errorLogger);
+		const scanner = getTestScanner([ SFVChecker ]);
 		await scanner.scanPath(scanPath);
 
 		expect(scanner.errors.count('file_missing')).toEqual(3);